Always handle MessageNoEagerSend before MessageTaskFinished
The MessageHandler passes some of its work to the ClusterWorkers. But it has to
be careful to ensure that all the MessageReleaseAccesses are handled before
MessageTaskFinished or MessageReleaseAccess (and finish). Otherwise there could
be a use-after-free. This commit also ensures that all the MessageNoEagerSends
are handled before deleting the task. Previously there was a (rare)
use-after-free if the MessageNoEagerSend is handled after deleting the task. As
a side-effect, all of the regions listed in a MessageNoEagerSend must be for
the same task. Add an assertion to verify that this is indeed the case.