checkDataRelease: turn conditional into an assertion
The comment saying that checkDataRelease needs to check task->isRemoteTask() to
avoid releasing for "taskwait on"s was misleading. A taskwait on is just an
if(0) task, which cannot be offloaded, so it never gets a DataReleaseStep.
Anyway, there is now an assertion to check that checkDataRelease is indeed only
called for remote tasks.