Fix misleading condition and comment in checkDataRelease
Taskwaits never release back to the offloader, but it was not clear from the
code. In fact, there was a condition "access->getObjectType() == taskwait_type"
in checkDataRelease with the comment that it releases top-level sinks. The
comment was wrong because the check matches taskwaits not top-level sinks
(which are of different type top_level_sink_type). But no such check is needed
in either case because the "or" condition matches: taskwaits and top-level
sinks don't have subaccesses.  The checkDataRelease function sometimes returns
true (release it) for taskwaits, but the return value is effectively ignored
because the only time this happens is as part of a debug check in
unfragmentTaskAccesses, which calls DataAccessStatusEffects, but doesn't act on
it. We make explicit in the code that taskwaits in fact never release.