Make namespace to process some other messages.
* src/cluster/NodeNamespace.hpp (_messageAction) : New variable to
contain an action message.  Use a simple pointer because the only one
message of this kind may be in queue at the time. We don't use the same
queue because the action messages need to be processed at the end.
(enqueueMessage) : Enqueue all the messages now.
(bodyPrivate) : Simplify code for lock taking and handle conditions.
* src/cluster/messages/Message.hpp (handleMessageNamespace) : New
optional virtual function to handle the message in the namespace; this
simplifies some code, but also will be needed in malleability to solve
the issue with the polling services.
* src/cluster/messages/MessageTaskNew.[hc]pp (enqueueMessage) :
Implement to call TaskOffloading.
* src/cluster/messages/MessageTaskNew.[hc]pp (handleMessage) : Use the
new function.
(handleMessageNamespace) : Call remoteTaskCreateAndSubmit in the
implementation.