Unify handleMessage for sysfinish using the same code in all nodes.
This will be useful to handle the sysFinish in the namespace and solve
the issues we had in the shutdown.

This also simplifies the code as we don't support working without
namespaces anymore.
* Makefile.am : Remove entry for MessageSysFinish.cpp.
* src/cluster/ClusterManager.[hc]pp (finishClusterNamespace) : Removed function.
(shutdownPhase1) : Use message handler that now contains the code in
finishClusterNamespace.
* src/cluster/messages/MessageSysFinish.cpp : Removed file, code moved
to header as it is very simple.
* src/cluster/messages/MessageSysFinish.hpp : Add code that was in cpp.