Workflow executeTask does delayed ops before starting workflow
After creating all the data copy steps in executeWorkflow (with the lock on the
task data accesses), there is a call to process all the dependency system
delayed operations: passing the namespace info and location, if found by
WriteID, to the immediate successor.  This call used to be after starting the
workflow, so very rarely (e.g. in namespacenewfail14), the workflow would
finish and remove the task and its successor before passing the namespace info.
It caused a rare use-after-free.  It is a bit dangerous that the calculation of
_isRemovable in the dependency system doesn't take into account whether the
namespace predecessor information has been received, but it does not.