Deactivate assertion that Taskfor doesn't change socket
If a taskfor is preempted by a higher-priority task or taskfor, it may be
rescheduled on a different TaskforGroup (socket). If that happens, the number
of CPUs, excluding the LeaderThread, may differ. The default chunksize is
calculated to make the work neatly fit among the original number of CPUs. After
rescheduling on a different socket, there may be some load imbalance. But
preemption implies multiple simultaneous taskfors, so hopefully it's not a
major problem. I don't know how to fix it properly because the code doesn't
support changing the chunksize mid-way, and constraining the scheduler to not
migrate taskfors could be even worse. This fixes taskfor/test3_taskfor14.c