Support namespace in-to-in propagation
The check in replaceMatchingInBottomMapLinkAndPropagate, as part of
linkTaskAccesses, which was supposed to disallow in-to-inout namespace
propagation also disallowed in-to-in namespace propagation. It actually only
allowed inout-to-inout propagation. Reenable in-to-in propagation, which
generally worked, and fix a couple of issues: (1) the connection of in accesses
can be done in any order at the remote end but only if the tasks are siblings,
and (2) if offloaded task T with inout has child A (with in) and successor B
(with in), then both A and B used to get T as their namespace predecessor.  But
if A (in) and B (in) are connected in the wrong order at the remote side there
would be a deadlock, since T has inout so won't pass satisfiability to B until
A completes (after B completes).  We therefore need to distinguish whether
namespace propagation is as a successor of a task or a child of it. Introduce a
second OffloadedTaskId for namespace propagation of the child tasks to refer to
their parent.