Fix hang for bottom map subtasks of lmalloc'ed regions
There used to be a hang after main finishes if a non-offloaded task allocates
data, e.g. with nanos6_lmalloc, and then finishes with a task on the bottom map
that accesses the allocated region. The task on the bottom map has the "on
bottom map" status bit set and, since there is no proper parent access, it will
never have a next. So it never becomes removable. It was OK if the task had the
"wait" clause, or it was an offloaded task (with wait, autowait or nowait, the
latter because of the top-level sink). In all these cases there would be a
taskwait or top-level sink as the next access and everything would be cleaned
up in handleExitTaskwait. But this did not happen if the parent task didn't
have wait and wasn't offloaded. This fixes nasty_test3_hang2 and
nasty_test3_hang3.