Don't disable release for unused parts of lmalloc'ed region
When some, but not all, parts of a locally lmalloc'ed region are not used by
any subtasks and the task did not have the "wait" clause, then the unused parts
never got released, causing a hang. This commit adds a missing condition when
creating the corresponding taskwait fragments.

This commit fixes auto_auto_true39.