Don't update location of local array when R/W satisfiability
A task may call nanos6_lmalloc before read satisfiability has arrived (from
an earlier task with auto). In that case it does not block, so when read
satisfiability arrives, it used to overwrite the correct location with the
old location. Add a condition to keep the existing location when read
satisfiability arrives for a strong local access.

This fixes auto/auto_true40.