registerLocalAccess updates fragment locations if any
When a task does a nanos6_lmalloc inside a task with auto, there may already be
a fragment covering that region. This would happen if the new region is inside
a (large) auto access that was already fragmented by a previous subtask. When that
happens, all fragments in the (large) access get created, even the ones for the
region that is being registered now. This means that registerLocalAccess needs to
check the fragments too. This fixes auto/auto_true41.