Never erase memory on setNUMAAffinity
The function checkAllocationCorrectness assumed that it is used only on
new allocations, but we use it also on distributed memory reallocation.
Such function erases the memory calling memset to check that the
allocation took place (for debugging purposes).
We move the call to the alloc function and not use it for distributed
memory because when redistributing allocations we need to preserve the
data.