Remove incorrect NUMA assertion for nanos6_dfree
The function NUMAManager::unsetNUMAAffinity used to have an assertion to check
that the region is registered with the NUMA directory (because why remove a
region unless it has already been registered?). The problem is that this
assertion doesn't work well with nanos6_dfree. nanos6_dmalloc only registers
NUMA regions for the parts of the dmalloc with home the current node, and it
only registers complete pages on the node. So it is possible that a dmalloc of
size less than 4 KB per node has no NUMA regions registered for the current
node. We could complicate nanos6_dfree to figure out exactly which regions were
registered on this node, and then have to make "dfree" versions of every
distribution policy (once we have more than one). But it is sufficient to just
remove the assertion and let unsetNUMAAffinity silently do nothing if there is
no NUMA region registered. This fixes #14.