Fix when mpi_ub_tag is not power-of-two minus one
The MPI messenger uses the MPI message tag to encode the MessageType (in the
bottom eight bits) and the message's unique ID (in the remainder of the
available bits).  We have to ensure that the message tag doesn't exceed the
value of mpi_ub_tag from the MPI library. This is done in createTag, which used
to do a bitwise AND with mpi_ub_tag. Unfortunately, it doesn't work if
mpi_ub_tag isn't a power-of-two minus one.  On CTE-POWER, for instance,
mpi_ub_tag equals 0x40000000, so createTag used to clear most of the bits with
the effect of setting the MessageType to zero.  Now, we use a bitmask equal to
the largest power-of-two minus one that doesn't exceed mpi_ub_tag.