Attempt to fix properly the Omar's issue with namespace shutdown
This is the right fix that doesn't start the namespace when there are
not more nodes and they will not be added in the future because it is
disabled at config level.

We can't rely on inClusterMode because it is false also when the
communicator is enabled but they can be added latter. In that case the
namespace needs to be added at the beginning "just in case".

However, when there is not communicator we can be sure that this won't
happen, so we can skip it completely.

Any error related to not enabling the namespace inn this condition
needs to be fixed as they can be cosidered bugs for untested
conditions.