Fix nanos6_cluster_device ComputePlace with monitoring=true
The ComputePlace constructor preallocates a taskfor for it, except when the
compute place is another cluster node; i.e. it is of type
nanos6_cluster_device. In this case, the preallocated taskfor becomes nullptr,
but the previous code was still calling Monitoring::taskCreated on it. This led
to an assertion failure or segmentation fault when monitoring was enabled.