MPI + OmpSs-2@Cluster: disable NUMA in dmalloc
NUMA support doesn't currently work when in MPI+OmpSs-2@Cluster mode.  It seems
to be because as the number of cores per instance changes it cannot unset the
NUMA affiliation for pages that have already been accessed. So later when the
distribution causes an in-use page to change NUMA node there is an error.