TaskloopOffloaders have weak accesses
The original TaskloopSource is split into one TaskloopOffloader per node, each
of which creates the TaskloopExecutors for that node. A TaskloopOffloader
should be a kind of TaskloopSource, so that all its accesses become weak (this
is done by register_access in RegisterDependencies.cpp).  Unfortunately
isTaskloopSource was not returning true at this point. So update
setTaskloopOffloader to ensure all TaskloopOffloaders are also TaskloopSources.