TaskloopOffloaders: don't nest "add task" state
The previous version created and submitted the TaskloopOffloaders inside
AddTask::submitTask, before leaving the "add task" runtime state (by calling
TrackingPoints::exitSubmitTask). It was therefore nesting the "add task"
runtime state (for a TaskloopOffloader) inside "add task" (for the
TaskloopSource), which is not normal behaviour. It's not clear anyone would
expect such a thing to happen. For instance, it caused an assertion failure in
debug mode when monitoring.enabled = true and "runtimestate" is enabled by
monitoring.areas. Avoid this behaviour by calling
TrackingPoints::exitSubmitTask before creating the TaskloopOffloaders.