Fix MPI + OmpSs-2@Cluster Extrae communication ranks
This fixes 'Update Extrae for MPI + OmpSs-2@Cluster', which did not lookup the
instrumentation rank of the communication partners on receives. It only did the
proper translation for receives, meaning that there were unmatched
communications that did not appear in the generated trace. This meant that some
of the communication lines were missing when tracing in MPI + OmpSs-2@Cluster
mode.