DLB global policy: Set correct initial #allocated CPUs
In DLB mode, it used to set the initial allocated number of cores (target
number to own) to be 1 on the master and 0 on the helpers. When using the
global allocation, this meant that the beginning of the execution was
unnecessarily slow until the global allocation policy runs for the first time.