Fix dmalloc distribution (non-DLB)
In DLB mode, the dmallocs are distributed over all the cores, which means that
an exact dmalloc distribution is only possible if the number of elements is an
exact multiple of the total number of cores across the apprank. The same thing
was happening outside DLB mode. Fix it so that without DLB (either when built
without DLB or built with DLB but without MPI+OmpSs-2@Cluster), the dmallocs
are distributed over the nodes. An exact distribution is therefore possible
whenever the number of elements is a multiple of the number of nodes. Do this
by "faking" an allocation of one core per node for the purpose of dmalloc
distribution.

This fixes manual/dmalloc_distribution.c