Fix Extrae function name for spawned functions (e.g. main)
The function name for "main" was (often) showing up as "Unresolved" in Paraver.
In general, the function name was "Unresolved" when returning to a spawned
function after it unblocks (which frequently happens for main).  The
constructor for Extrae::user_fct_t sets the Extrae value to be registered via
ExtraeAPI::register_function_address, which is used for the
RUNNING_FUNCTION_NAME and RUNNING_CODE_LOCATION events. For normal tasks, the
value is the function pointer and for spawned tasks, it is the address of the
taskInfo.  Unfortunately some of the places that emit these events (e.g.
returnTask) always use the function pointer, so they did not handle the special
case for spawned functions.  This issue seems to have been introduced by "Fix
memory issue in extrae instrumentation.".