Fix WriteID bottleneck when unfragmenting lots of taskwaits
When there are a huge number of subtasks, unfragmentTaskwaits was taking a long
time unfragmenting the taskwaits before removing them. The bottleneck was in
registering new WriteIDs, which is needed when the merged accesses didn't have
the same WriteIDs. In this case a new WriteID is allocated for the merged
taskwait. This used to be done one access at a time; i.e. each time a taskwait
is merged into the previous one.  Now it registers in WriteID just once
irrespective of the number of accesses that are eventually merged into one.
This fixed a performance issue for synthetic_unbalanced from the DLB benchmarks
when using 50000 tasks.